requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.3.2  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==21.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.1.2  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-beat
flower==1.0.0  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.1.13  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.7.0  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.45.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-redis==5.0.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.8.0 # https://github.com/adamchainz/django-cors-headers

# Admin
# ------------------------------------------------------------------------------
django-jet2==1.0.12
FeedParser==5.2.1
django-ipware==3.0.1
django-autocomplete-light==3.8.2
# django-flatpages-tinymce==0.1.1
django-tinymce==3.3.0
django-filebrowser-no-grappelli==3.8.0
google-api-python-client==1.4.1
django-admin-honeypot==1.1.0
django-cookie-law==2.0.5  # https://pypi.org/project/django-cookie-law/

# List of countries
# ------------------------------------------------------------------------------
# django-countries-plus==1.3.1


# Fileifeld upload validator
python-magic==0.4.24

# Add inclusion tags easily
django-classy-tags==2.0.0 

# Minify HTML in django
# ------------------------------------------------------------------------------
django-shell-plus==1.1.7

# Minify HTML in django
# ------------------------------------------------------------------------------
# Todo: find the uses of these to minify wesbites and others to improve security
django-htmlmin==0.11.0
python-memcached==1.59


# API NECESSARY APPS
# ------------------------------------------------------------------------------
djangorestframework-simplejwt==4.6.0 # Add proper token creation url links for app
drf-yasg==1.20.0 # Add documentation for api

# Add swagger for friendlier api view
# ------------------------------------------------------------------------------
django-rest-swagger==2.2.0

# Documentation API
# ------------------------------------------------------------------------------
coreapi==2.3.3 # https://github.com/core-api/python-client

# Add image resize for models
# ------------------------------------------------------------------------------
django-resized==0.3.11


# Add medium type of read time to a post
# ------------------------------------------------------------------------------
readtime==1.1.1
django-comments-dab==2.6.0



# Django Webpack for smoother coding
django-webpack-loader==1.3.0 
python-webpack-boilerplate==0.0.3 

# Dynamic Scraper for django
# ------------------------------------------------------------------------------
# Scrapy is a fast high-level web crawling and web scraping framework, 
# used to crawl websites and extract structured data from their pages. 
# It can be used for a wide range of purposes, from data mining to 
# monitoring and automated testing.
# ------------------------------------------------------------------------------
# webdriver-manager==3.4.2
# lxml==4.6.3 
# scrapy-djangoitem==1.1.1
# word2word==1.0.0 
# requests-html==0.10.0

#Todo: Django Filter Search, Ajax Commenting, Language Integration, Cookie Consent
django-category==2.0.1 #
django-csp==3.7

# convert html to pdf and filter
# xhtml2pdf==0.2.5
django-filter==2.4.0 

# Translation with google translate
django-google-translate==1.1